Move common dataclasses to free energy
